Approximate sine waves with a perceptron.
The network has two inputs. The first is for frequency; the second is
for time. It has one output, which is roughly the value of the sine
function at the given time with the given frequency. In other words, if
the inputs of the network are `freq` and `x`, its output approximates
`sin(freq * x)`.

You can relate `x` to time in seconds (`t`) with this equation:
`x = 2pi * t`